+++ js/game.js
@@ -5,7 +5,7 @@ var NaNalert = false;
 var gameEnded = false;
 
 let VERSION = {
-	num: "1.3",
+	num: "1.3.1",
 	name: "Tabception... ception!"
 }
 
@@ -231,21 +231,21 @@ function respecBuyables(layer) {
 }
 
 function canAffordUpg(layer, id) {
-	upg = layers[layer].upgrades[id]
-	cost = tmp.upgrades[layer][id].cost
+	let upg = layers[layer].upgrades[id]
+	let cost = tmp.upgrades[layer][id].cost
 	return canAffordPurchase(layer, upg, cost) 
 }
 
 function hasUpg(layer, id){
-	return (player[layer].upgrades.includes(toNumber(id)))
+	return (player[layer].upgrades.includes(toNumber(id)) || player[layer].upgrades.includes(id.toString()))
 }
 
 function hasMilestone(layer, id){
-	return (player[layer].milestones.includes(toNumber(id)) || player[layer].milestones.includes(id))
+	return (player[layer].milestones.includes(toNumber(id)) || player[layer].milestones.includes(id.toString()))
 }
 
 function hasChall(layer, id){
-	return (player[layer].challs.includes(toNumber(id)))
+	return (player[layer].challs.includes(toNumber(id)) || player[layer].challs.includes(id.toString()))
 }
 
 function upgEffect(layer, id){
@@ -280,8 +280,8 @@ function buyUpg(layer, id) {
 	if (!player[layer].unl) return
 	if (!layers[layer].upgrades[id].unl()) return
 	if (player[layer].upgrades.includes(id)) return
-	upg = layers[layer].upgrades[id]
-	cost = tmp.upgrades[layer][id].cost
+	let upg = layers[layer].upgrades[id]
+	let cost = tmp.upgrades[layer][id].cost
 
 	if (upg.currencyInternalName){
 		let name = upg.currencyInternalName
@@ -329,14 +329,17 @@ function resetRow(row) {
 }
 
 function startChall(layer, x) {
+	let enter = false
 	if (!player[layer].unl) return
 	if (player[layer].active == x) {
 		completeChall(layer, x)
 		delete player[layer].active
 	} else {
-		player[layer].active = x
-	}
+		enter = true
+	}	
 	doReset(layer, true)
+	if(enter) player[layer].active = x
+
 	updateChallTemp(layer)
 }